requirements/base.txt
asgiref==3.6.0
boto3==1.26.59
celery==5.3.0
Django==4.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-ses==3.5.0
django-redis==5.2.0
PyJWT==2.6.0
python-environ==0.4.54
pytz==2023.3
sqlparse==0.4.3
psycopg2-binary==2.9.5
drf-yasg==1.21.5
django-storages==1.13.2
django-parler==2.3
Pillow==9.5.0
redis==4.4.2
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
